+++ packages/transform-geojson/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@nteract/transform-geojson",
+  "version": "1.0.0",
+  "description": "GeoJSON Transform",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "BSD-3-Clause"
+}
