+++ packages/transform-model-debug/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@nteract/tranform-model-debugtransform-vega",
+  "version": "1.0.0",
+  "description": "Transform for debug usage with nteract models",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "peerDependencies": {
+    "react": "^15.4.2"
+  },
+  "devDependencies": {
+    "react": "^15.4.2"
+  },
+  "license": "BSD-3-Clause"
+}
