+++ packages/transform-vega/index.js
