+++ packages/transform-vega/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@nteract/transform-vega",
+  "version": "1.0.0",
+  "description": "Vega Transform",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "peerDependencies": {
+    "react": "^15.4.2"
+  },
+  "devDependencies": {
+    "react": "^15.4.2"
+  },
+  "license": "BSD-3-Clause"
+}
